Retrieve Census data from a given API

Retrieve Census data from a given API

getCensus(name, vintage = NULL, key = Sys.getenv("CENSUS_KEY"), vars,
  region = NULL, regionin = NULL, time = NULL, date = NULL,
  period = NULL, monthly = NULL, category_code = NULL,
  data_type_code = NULL, naics = NULL, pscode = NULL,
  naics2012 = NULL, naics2007 = NULL, naics2002 = NULL,
  naics1997 = NULL, sic = NULL)

Arguments

name
API name - e.g. acs5. See list at https://api.census.gov/data.html

vintage
Year of dataset, e.g. 2014 - not required for timeseries APIs

key
Your Census API key, gotten from https://api.census.gov/data/key_signup.html

vars
List of variables to get

region
Geography to get

regionin
Optional hierarchical geography to limit region

time, date, period, monthly
Optional arguments used for some time series APIs

category_code, data_type_code
Arguments used in Economic Indicators Time Series API

naics, pscode
Arguments used in Annual Survey of Manufactures API

naics2012, naics2007, naics2002, naics1997, sic
Arguments used in Economy Wide Key Statistics APIs and Business Patterns APIs

Examples

df <- getCensus(name = "acs/acs5", vintage = 2017,
        vars = c("B01001_001E", "NAME", "B01002_001E", "B19013_001E"),
        region = "tract:*", regionin = "state:06")
head(df)

#>   state county  tract B01001_001E                                          NAME
#> 1    06    001 400100        2991 Census Tract 4001, Alameda County, California
#> 2    06    001 400200        1997 Census Tract 4002, Alameda County, California
#> 3    06    001 400300        5123 Census Tract 4003, Alameda County, California
#> 4    06    001 400400        3991 Census Tract 4004, Alameda County, California
#> 5    06    001 400500        3944 Census Tract 4005, Alameda County, California
#> 6    06    001 400600        1635 Census Tract 4006, Alameda County, California
#>   B01002_001E B19013_001E
#> 1        53.7      208393
#> 2        43.1      147500
#> 3        42.4       88173
#> 4        35.9      102821
#> 5        32.7       92375
#> 6        36.2       86458

# Use American Community Survey variable groups to get all data from a given table.
# This returns estimates as well as margins of error and annotation flags.
acs_group <- getCensus(name = "acs/acs5", vintage = 2017,
        vars = c("NAME", "group(B19013)"),
        region = "county:*")

#> Warning: NAs introduced by coercion

        head(acs_group)

#>   state county                                 NAME         GEO_ID B19013_001E
#> 1    72    047       Corozal Municipio, Puerto Rico 0500000US72047       14752
#> 2    72    095       Maunabo Municipio, Puerto Rico 0500000US72095       17636
#> 3    72    111      Peñuelas Municipio, Puerto Rico 0500000US72111       16868
#> 4    72    113         Ponce Municipio, Puerto Rico 0500000US72113       16561
#> 5    72    131 San Sebastián Municipio, Puerto Rico 0500000US72131       14275
#> 6    72    135      Toa Alta Municipio, Puerto Rico 0500000US72135       29672
#>   B19013_001M NAME_1 B19013_001EA B19013_001MA
#> 1        1416     NA         <NA>         <NA>
#> 2        3292     NA         <NA>         <NA>
#> 3        1444     NA         <NA>         <NA>
#> 4         594     NA         <NA>         <NA>
#> 5        1197     NA         <NA>         <NA>
#> 6        1584     NA         <NA>         <NA>

# Retreive block-level data within a specific state and county using a nested regionin argument
data2010 <- getCensus(name = "dec/sf1", vintage = 2010,
vars = c("P001001", "H010001"),
region = "block:*", regionin = "state:36+county:027")
head(data2010)

#>   state county  tract block P001001 H010001
#> 1    36    027 010000  1000      31      31
#> 2    36    027 010000  1011      17      17
#> 3    36    027 010000  1028      41      41
#> 4    36    027 010000  1001       0       0
#> 5    36    027 010000  1031       0       0
#> 6    36    027 010000  1002       4       4

# Retreive block-level data for Decennial Census sf1, 2000
# Note, for this dataset a tract needs to be specified to retrieve blocks
data2000 <- getCensus(name = "sf1", vintage = 2000,
        vars = c("P001001", "P003001"),
region = "block:*", regionin = "state:36+county:27+tract:010000")
head(data2000)

#>   state county  tract block P001001 P003001
#> 1    36    027 010000  1000      18      18
#> 2    36    027 010000  1001      26      26
#> 3    36    027 010000  1002      59      59
#> 4    36    027 010000  1003      67      67
#> 5    36    027 010000  1004      52      52
#> 6    36    027 010000  1005     116     116

# Get time series data
saipe <- getCensus(name = "timeseries/poverty/saipe",
        vars = c("NAME", "SAEPOVRT0_17_PT", "SAEPOVRTALL_PT"),
        region = "state:*", time = 2011)
head(saipe)

#>   time state                 NAME SAEPOVRT0_17_PT SAEPOVRTALL_PT
#> 1 2011    08             Colorado            17.7           13.4
#> 2 2011    09          Connecticut            14.6           10.8
#> 3 2011    10             Delaware            18.8           12.6
#> 4 2011    11 District of Columbia            30.9           19.1
#> 5 2011    12              Florida            25.1             17
#> 6 2011    01              Alabama            27.6           19.1

# Get county business patterns data for a specific NAICS sector
cbp_2016 <- getCensus(name = "cbp",
 vintage = "2016",
 vars = c("EMP", "ESTAB", "NAICS2012_TTL", "GEO_TTL"),
 region = "state:*",
 naics2012 = "23")
 head(cbp_2016)

#>   state    EMP ESTAB NAICS2012_TTL    GEO_TTL NAICS2012
#> 1    01  82327  7424  Construction    Alabama        23
#> 2    02  17022  2521  Construction     Alaska        23
#> 3    04 141740 11921  Construction    Arizona        23
#> 4    05  45609  5293  Construction   Arkansas        23
#> 5    06 723574 71981  Construction California        23
#> 6    08 152325 17730  Construction   Colorado        23