Retrieve Census data from a given API

Retrieve Census data from a given API

Usage

getCensus(
  name,
  vintage = NULL,
  key = Sys.getenv("CENSUS_KEY"),
  vars,
  region = NULL,
  regionin = NULL,
  time = NULL,
  year = NULL,
  date = NULL,
  period = NULL,
  monthly = NULL,
  show_call = FALSE,
  category_code = NULL,
  data_type_code = NULL,
  naics = NULL,
  pscode = NULL,
  naics2012 = NULL,
  naics2007 = NULL,
  naics2002 = NULL,
  naics1997 = NULL,
  sic = NULL,
  ...
)

Arguments

name

API name - e.g. acs5. See list at https://api.census.gov/data.html

vintage

Year of dataset, e.g. 2014 - not required for timeseries APIs

key

Your Census API key, gotten from https://api.census.gov/data/key_signup.html

vars

List of variables to get

region

Geography to get

regionin

Optional hierarchical geography to limit region

time, year, date, period, monthly

Optional arguments used for some time series APIs

show_call

List the underlying API call sent to the Census Bureau and other info

category_code, data_type_code

Arguments used in Economic Indicators Time Series API

naics, pscode

Arguments used in Annual Survey of Manufactures API

naics2012, naics2007, naics2002, naics1997, sic

Arguments used in Economy Wide Key Statistics APIs and Business Patterns APIs

...

Other valid parameters to pass to the Census API. Note: the APIs are case sensitive.

Examples

df <- getCensus(name = "acs/acs5", vintage = 2017,
vars = c("B01001_001E", "NAME", "B01002_001E", "B19013_001E"),
region = "tract:*", regionin = "state:06")
head(df)
#>   state county  tract B01001_001E
#> 1    06    037 503502        4359
#> 2    06    037 504102           7
#> 3    06    037 265700        5181
#> 4    06    037 267300        4751
#> 5    06    037 267502        3524
#> 6    06    037 267700        3141
#>                                                   NAME  B01002_001E B19013_001E
#> 1 Census Tract 5035.02, Los Angeles County, California         35.8       73979
#> 2 Census Tract 5041.02, Los Angeles County, California -666666666.0  -666666666
#> 3    Census Tract 2657, Los Angeles County, California         47.5       84327
#> 4    Census Tract 2673, Los Angeles County, California         34.6       61290
#> 5 Census Tract 2675.02, Los Angeles County, California         30.8       80313
#> 6    Census Tract 2677, Los Angeles County, California         35.0       84608

# Use American Community Survey variable groups to get all data from a given table.
# This returns estimates as well as margins of error and annotation flags.
acs_group <- getCensus(name = "acs/acs5",
vintage = 2017,
vars = c("NAME", "group(B19013)"),
region = "county:*")
#> Warning: NAs introduced by coercion
head(acs_group)
#>   state county                          NAME B19013_001E B19013_001EA
#> 1    30    019       Daniels County, Montana       53591         <NA>
#> 2    72    073 Jayuya Municipio, Puerto Rico       15615         <NA>
#> 3    30    079       Prairie County, Montana       50850         <NA>
#> 4    39    173             Wood County, Ohio       58033         <NA>
#> 5    24    023      Garrett County, Maryland       48174         <NA>
#> 6    08    011         Bent County, Colorado       32500         <NA>
#>   B19013_001M B19013_001MA         GEO_ID NAME_1
#> 1        3318         <NA> 0500000US30019     NA
#> 2        3424         <NA> 0500000US72073     NA
#> 3        6481         <NA> 0500000US30079     NA
#> 4        1604         <NA> 0500000US39173     NA
#> 5        2234         <NA> 0500000US24023     NA
#> 6        3442         <NA> 0500000US08011     NA

# Retreive block-level data within a specific tract using a nested regionin argument
data2010 <- getCensus(name = "dec/sf1",
vintage = 2010,
vars = c("NAME","P001001"),
region = "block:*",
regionin = "state:36+county:027+tract:010000")
head(data2010)
#>   state county  tract block
#> 1    36    027 010000  1000
#> 2    36    027 010000  1011
#> 3    36    027 010000  1028
#> 4    36    027 010000  1001
#> 5    36    027 010000  1031
#> 6    36    027 010000  1002
#>                                                                     NAME
#> 1 Block 1000, Block Group 1, Census Tract 100, Dutchess County, New York
#> 2 Block 1011, Block Group 1, Census Tract 100, Dutchess County, New York
#> 3 Block 1028, Block Group 1, Census Tract 100, Dutchess County, New York
#> 4 Block 1001, Block Group 1, Census Tract 100, Dutchess County, New York
#> 5 Block 1031, Block Group 1, Census Tract 100, Dutchess County, New York
#> 6 Block 1002, Block Group 1, Census Tract 100, Dutchess County, New York
#>   P001001
#> 1      31
#> 2      17
#> 3      41
#> 4       0
#> 5       0
#> 6       4

# Get poverty rates for children and all ages over time
saipe <- getCensus(name = "timeseries/poverty/saipe",
vars = c("NAME", "SAEPOVRT0_17_PT", "SAEPOVRTALL_PT"),
region = "state:01",
year = "2000:2019")
head(saipe)
#>   state    NAME SAEPOVRT0_17_PT SAEPOVRTALL_PT YEAR
#> 1    01 Alabama            20.5           14.6 2000
#> 2    01 Alabama            22.1           15.7 2001
#> 3    01 Alabama            21.6           15.4 2002
#> 4    01 Alabama            22.3           15.3 2003
#> 5    01 Alabama            22.6           16.1 2004
#> 6    01 Alabama            24.3           16.9 2005

# Get county business patterns data for a specific NAICS sector
cbp_2016 <- getCensus(name = "cbp",
vintage = "2016",
vars = c("EMP", "ESTAB", "NAICS2012_TTL", "GEO_TTL"),
region = "state:*",
naics2012 = "23")
head(cbp_2016)
#>   state    EMP ESTAB NAICS2012_TTL     GEO_TTL NAICS2012
#> 1    27 118033 16360  Construction   Minnesota        23
#> 2    28  44551  3796  Construction Mississippi        23
#> 3    29 121494 13263  Construction    Missouri        23
#> 4    30  25036  5326  Construction     Montana        23
#> 5    31  45909  6357  Construction    Nebraska        23
#> 6    32  72817  4839  Construction      Nevada        23