Retrieve Census data from a given API

Retrieve Census data from a given API

getCensus(name, vintage = NULL, key = Sys.getenv("CENSUS_KEY"), vars,
  region, regionin = NULL, time = NULL, date = NULL, period = NULL,
  monthly = NULL, category_code = NULL, data_type_code = NULL,
  naics = NULL, pscode = NULL, naics2012 = NULL, naics2007 = NULL,
  naics2002 = NULL, naics1997 = NULL, sic = NULL)

Arguments

name
API name - e.g. acs5. See list at https://api.census.gov/data.html

vintage
Year of dataset, e.g. 2014 - not required for timeseries APIs

key
Your Census API key, gotten from https://api.census.gov/data/key_signup.html

vars
List of variables to get

region
Geography to get

regionin
Optional hierarchical geography to limit region

time, date, period, monthly
Optional arguments used for some time series APIs

category_code, data_type_code
Arguments used in Economic Indicators Time Series API

naics, pscode
Arguments used in Annual Survey of Manufactures API

naics2012, naics2007, naics2002, naics1997, sic
Arguments used in Economy Wide Key Statistics APIs and Business Patterns APIs

Examples

df <- getCensus(name = "acs/acs5", vintage = 2016,
        vars = c("B01001_001E", "NAME", "B01002_001E", "B19013_001E"),
        region = "tract:*", regionin = "state:06")
head(df)

#>   state county  tract B01001_001E                                          NAME
#> 1    06    001 400100        3018 Census Tract 4001, Alameda County, California
#> 2    06    001 400200        1960 Census Tract 4002, Alameda County, California
#> 3    06    001 400300        5236 Census Tract 4003, Alameda County, California
#> 4    06    001 400400        4171 Census Tract 4004, Alameda County, California
#> 5    06    001 400500        3748 Census Tract 4005, Alameda County, California
#> 6    06    001 400600        1661 Census Tract 4006, Alameda County, California
#>   B01002_001E B19013_001E
#> 1        54.2      177417
#> 2        43.6      153125
#> 3        41.3       85313
#> 4        36.6       99539
#> 5        34.1       83650
#> 6        34.8       61597

# Use American Community Survey variable groups to get all data from a given table.
# This returns estimates as well as margins of error and annotation flags.
acs_group <- getCensus(name = "acs/acs5", vintage = 2016,
        vars = c("NAME", "group(B19013)"),
        region = "county:*")
        head(acs_group)

#>   state county                    NAME B19013_001E B19013_001M B19013_001M_1
#> 1    01    001 Autauga County, Alabama       53099        2631          2631
#> 2    01    003 Baldwin County, Alabama       51365         991           991
#> 3    01    005 Barbour County, Alabama       33956        2655          2655
#> 4    01    007    Bibb County, Alabama       39776        3306          3306
#> 5    01    009  Blount County, Alabama       46212        2443          2443
#> 6    01    011 Bullock County, Alabama       29335        5435          5435
#>   B19013_001EA B19013_001MA
#> 1         <NA>         <NA>
#> 2         <NA>         <NA>
#> 3         <NA>         <NA>
#> 4         <NA>         <NA>
#> 5         <NA>         <NA>
#> 6         <NA>         <NA>

# Retreive block-level data within a specific state and county using a nested regionin argument
data2010 <- getCensus(name = "sf1", vintage = 2010,
vars = c("P0010001", "P0030001"),
region = "block:*", regionin = "state:36+county:27")
head(data2010)

#>   state county  tract block P0010001 P0030001
#> 1    36    027 010000  1020       73       73
#> 2    36    027 010000  1023        0        0
#> 3    36    027 010000  1030       68       68
#> 4    36    027 010000  1031        0        0
#> 5    36    027 010000  1032        0        0
#> 6    36    027 010000  1033        0        0

# Retreive block-level data for Decennial Census sf1, 2000
# Note, for this dataset a tract needs to be specified to retrieve blocks
data2000 <- getCensus(name = "sf1", vintage = 2000,
        vars = c("P001001", "P003001"),
region = "block:*", regionin = "state:36+county:27+tract:010000")
head(data2000)

#>   state county  tract block P001001 P003001
#> 1    36    027 010000  1000      18      18
#> 2    36    027 010000  1001      26      26
#> 3    36    027 010000  1002      59      59
#> 4    36    027 010000  1003      67      67
#> 5    36    027 010000  1004      52      52
#> 6    36    027 010000  1005     116     116

# Get time series data
saipe <- getCensus(name = "timeseries/poverty/saipe",
        vars = c("NAME", "SAEPOVRT0_17_PT", "SAEPOVRTALL_PT"),
        region = "state:*", time = 2011)
head(saipe)

#>   time state       NAME SAEPOVRT0_17_PT SAEPOVRTALL_PT
#> 1 2011    01    Alabama            27.6           19.1
#> 2 2011    02     Alaska            14.7           10.8
#> 3 2011    04    Arizona            27.2             19
#> 4 2011    05   Arkansas            27.8           19.3
#> 5 2011    06 California            22.8           16.6
#> 6 2011    08   Colorado            17.7           13.4

# Get county business patterns data for a specific NAICS sector
cbp_2016 <- getCensus(name = "cbp",
 vintage = "2016",
 vars = c("EMP", "ESTAB", "NAICS2012_TTL", "GEO_TTL"),
 region = "state:*",
 naics2012 = "23")

#> Error in apiParse(req): The Census Bureau returned the following error message:
#>  There was an error while running your query.  We've logged the error and we'll correct it ASAP.  Sorry for the inconvenience.

 head(cbp_2016)

#> Error in head(cbp_2016): object 'cbp_2016' not found