Get dataset metadata on all available APIs as a data frame

Scrapes https://api.census.gov/data.json and returns a dataframe that includes: title, name, vintage (where applicable), url, isTimeseries (binary), temporal (helpful for some time series), description, modified date

listCensusApis()

Examples

apis <- listCensusApis()
head(apis)

#>                                                    title             name
#> 100                           ACS 1-Year Detailed Tables         acs/acs1
#> 70                              ACS 1-Year Data Profiles acs/acs1/profile
#> 23  2005-2009 American Community Survey: Migration Flows        acs/flows
#> 136 2006-2010 American Community Survey: Migration Flows        acs/flows
#> 72  2007-2011 American Community Survey: Migration Flows        acs/flows
#> 144 2008-2012 American Community Survey: Migration Flows        acs/flows
#>     vintage                                               url isTimeseries
#> 100    2015         https://api.census.gov/data/2015/acs/acs1           NA
#> 70     2015 https://api.census.gov/data/2015/acs/acs1/profile           NA
#> 23     2009        https://api.census.gov/data/2009/acs/flows           NA
#> 136    2010        https://api.census.gov/data/2010/acs/flows           NA
#> 72     2011        https://api.census.gov/data/2011/acs/flows           NA
#> 144    2012        https://api.census.gov/data/2012/acs/flows           NA
#>         temporal
#> 100 unidentified
#> 70  unidentified
#> 23     2006-2010
#> 136    2006-2010
#> 72     2006-2010
#> 144    2006-2010
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        description
#> 100     The American Community Survey (ACS) is a nationwide survey designed to provide communities a fresh look at how they are changing. The ACS replaced the decennial census long form in 2010 and thereafter by collecting long form type information throughout the decade rather than only once every 10 years. Questionnaires are mailed to a sample of addresses to obtain information about households -- that is, about each person and the housing unit itself. The American Community Survey produces demographic, social, housing and economic estimates in the form of 1 and 5-year estimates based on population thresholds. The strength of the ACS is in estimating population and housing characteristics. The 2012 data provide key estimates for each of the topic areas covered by the ACS for the nation, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more. Although the ACS produces population, demographic and housing unit estimates,it is the Census Bureau's Population Estimates Program that produces and disseminates the official estimates of the population for the nation, states, counties, cities and towns, and estimates of housing units for states and counties. For 2010 and other decennial census years, the Decennial Census provides the official counts of population and housing units.
#> 70  The American Community Survey (ACS) is a nationwide survey designed to provide communities a fresh look at how they are changing. The ACS replaced the decennial census long form in 2010 and thereafter by collecting long form type information throughout the decade rather than only once every 10 years. Questionnaires are mailed to a sample of addresses to obtain information about households -- that is, about each person and the housing unit itself. The American Community Survey produces demographic, social, housing and economic estimates in the form of 1 and 5-year estimates based on population thresholds. The strength of the ACS is in estimating population and housing characteristics. The data profiles provide key estimates for each of the topic areas covered by the ACS for the nation, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more. Although the ACS produces population, demographic and housing unit estimates,it is the Census Bureau's Population Estimates Program that produces and disseminates the official estimates of the population for the nation, states, counties, cities and towns, and estimates of housing units for states and counties. For 2010 and other decennial census years, the Decennial Census provides the official counts of population and housing units.
#> 23                                                                                                                                                                                                                                                                                                                  Migration flows are derived from the relationship between the location of current residence in the American Community Survey (ACS) sample and the responses given to the migration question "Where did you live 1 year ago?".  There are flow statistics (moved in, moved out, and net moved) between county or minor civil division (MCD) of residence and county, MCD, or world region of residence 1 year ago.  Estimates for MCDs are only available for the 12 strong-MCD states, where the MCDs have the same government functions as incorporated places.  Migration flows between metropolitan statistical areas are available starting with the 2009-2013 5-year ACS dataset.   Flow statistics are available by three or four variables for each dataset starting with the 2006-2010 5-year ACS datasets.  The variables change for each dataset and do not repeat in overlapping datasets.  In addition to the flow estimates, there are supplemental statistics files that contain migration/geographical mobility estimates (e.g., nonmovers, moved to a different state, moved from abroad) for each county, MCD, or metro area.
#> 136                                                                                                                                                                                                                                                                                                                 Migration flows are derived from the relationship between the location of current residence in the American Community Survey (ACS) sample and the responses given to the migration question "Where did you live 1 year ago?".  There are flow statistics (moved in, moved out, and net moved) between county or minor civil division (MCD) of residence and county, MCD, or world region of residence 1 year ago.  Estimates for MCDs are only available for the 12 strong-MCD states, where the MCDs have the same government functions as incorporated places.  Migration flows between metropolitan statistical areas are available starting with the 2009-2013 5-year ACS dataset.   Flow statistics are available by three or four variables for each dataset starting with the 2006-2010 5-year ACS datasets.  The variables change for each dataset and do not repeat in overlapping datasets.  In addition to the flow estimates, there are supplemental statistics files that contain migration/geographical mobility estimates (e.g., nonmovers, moved to a different state, moved from abroad) for each county, MCD, or metro area.
#> 72                                                                                                                                                                                                                                                                                                                  Migration flows are derived from the relationship between the location of current residence in the American Community Survey (ACS) sample and the responses given to the migration question "Where did you live 1 year ago?".  There are flow statistics (moved in, moved out, and net moved) between county or minor civil division (MCD) of residence and county, MCD, or world region of residence 1 year ago.  Estimates for MCDs are only available for the 12 strong-MCD states, where the MCDs have the same government functions as incorporated places.  Migration flows between metropolitan statistical areas are available starting with the 2009-2013 5-year ACS dataset.   Flow statistics are available by three or four variables for each dataset starting with the 2006-2010 5-year ACS datasets.  The variables change for each dataset and do not repeat in overlapping datasets.  In addition to the flow estimates, there are supplemental statistics files that contain migration/geographical mobility estimates (e.g., nonmovers, moved to a different state, moved from abroad) for each county, MCD, or metro area.
#> 144                                                                                                                                                                                                                                                                                                                 Migration flows are derived from the relationship between the location of current residence in the American Community Survey (ACS) sample and the responses given to the migration question "Where did you live 1 year ago?".  There are flow statistics (moved in, moved out, and net moved) between county or minor civil division (MCD) of residence and county, MCD, or world region of residence 1 year ago.  Estimates for MCDs are only available for the 12 strong-MCD states, where the MCDs have the same government functions as incorporated places.  Migration flows between metropolitan statistical areas are available starting with the 2009-2013 5-year ACS dataset.   Flow statistics are available by three or four variables for each dataset starting with the 2006-2010 5-year ACS datasets.  The variables change for each dataset and do not repeat in overlapping datasets.  In addition to the flow estimates, there are supplemental statistics files that contain migration/geographical mobility estimates (e.g., nonmovers, moved to a different state, moved from abroad) for each county, MCD, or metro area.
#>       modified
#> 100 2017-06-08
#> 70  2017-06-08
#> 23  2016-12-23
#> 136 2016-06-22
#> 72  2016-06-22
#> 144 2016-06-22