Get general information about available datasets

Scrapes https://api.census.gov/data.json and returns a dataframe that includes columns for dataset title, description, name, vintage, url, dataset type, and other useful fields.

Usage

listCensusApis(name = NULL, vintage = NULL)

Arguments

name

Optional complete or partial API dataset programmatic name. For example, "acs", "acs/acs5", "acs/acs5/subject". If using a partial name, this needs to be the left-most part of the dataset name before /, e.g. "timeseries/eits" or "dec" or "acs/acs5".

vintage

Optional vintage (year) of dataset.

Value

A data frame with the following columns:

title: Short written description of the dataset.

name: Programmatic name of the dataset.

vintage: Year of the survey, for use with microdata and aggregate datasets.

type: Dataset type, which is either "Aggregate", "Microdata", or "Timeseries".

temporal: Time period of the dataset. Warning: not always documented.

spatial: Spatial region of the dataset. Warning: not always documented.

url: Base URL of the dataset endpoint.

modified: Date last modified. Warning: sometimes out of date.

description: Long written description of the dataset.

contact: Email address for specific questions about the Census Bureau survey.

See also

Examples

# Get information about every dataset available in the APIs
apis <- listCensusApis()
head(apis)
#>                                                             title          name
#> 1                        Current Population Survey: Basic Monthly cps/basic/jan
#> 2 Current Population Survey Annual Social and Economic Supplement  cps/asec/mar
#> 3                        Current Population Survey: Basic Monthly cps/basic/apr
#> 4                        Current Population Survey: Basic Monthly cps/basic/aug
#> 5                        Current Population Survey: Basic Monthly cps/basic/dec
#> 6                        Current Population Survey: Basic Monthly cps/basic/feb
#>   vintage      type        temporal       spatial
#> 1    2025 Microdata 2025-01/2025-01 United States
#> 2    2024 Microdata 2024-03/2024-03 United States
#> 3    2024 Microdata 2024-04/2024-04 United States
#> 4    2024 Microdata 2024-08/2024-08 United States
#> 5    2024 Microdata 2024-12/2024-12 United States
#> 6    2024 Microdata 2024-02/2024-02 United States
#>                                             url              modified
#> 1 http://api.census.gov/data/2025/cps/basic/jan 2025-01-21 11:17:14.0
#> 2  http://api.census.gov/data/2024/cps/asec/mar 2024-07-30 10:09:36.0
#> 3 http://api.census.gov/data/2024/cps/basic/apr 2024-01-11 15:35:41.0
#> 4 http://api.census.gov/data/2024/cps/basic/aug 2024-01-11 15:35:41.0
#> 5 http://api.census.gov/data/2024/cps/basic/dec 2024-01-11 15:35:41.0
#> 6 http://api.census.gov/data/2024/cps/basic/feb 2024-01-11 15:35:41.0
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             description
#> 1                                                                                                                                                                                                To provide estimates of employment, unemployment, and other characteristics of the general labor force, of the population as a whole, and of various subgroups of the population. Monthly labor force data for the country are used by the Bureau of Labor Statistics (BLS) to determine the distribution of funds under the Job Training Partnership Act. These data are collected through combined computer-assisted personal interviewing (CAPI) and computer-assisted telephone interviewing (CATI). In addition to the labor force data, the CPS basic funding provides annual data on work experience, income, health insurance, and migration data from the Annual Social and Economic Supplement (ASEC), and on school enrollment of the population from the October Supplement. Other supplements, some of which are sponsored by other agencies, are conducted biennially or intermittently.
#> 2 The Annual Social and Economic Supplement or March CPS supplement is the primary source of detailed information on income and work experience in the United States. Numerous publications based on this survey are issued each year by the Bureaus of Labor Statistics and Census. A public-use microdata file is available for private researchers, who also produce many academic and policy-related documents based on these data. The Annual Social and Economic Supplement is used to generate the annual Population Profile of the United States, reports on geographical mobility and educational attainment, and detailed analysis of money income and poverty status. The labor force and work experience data from this survey are used to profile the U.S. labor market and to make employment projections. To allow for the same type of in-depth analysis of hispanics, additional Hispanic sample units are added to the basic CPS sample in March each year. Additional weighting is also performed so that estimates can be made for households and families, in addition to persons.
#> 3                                                                                                                                                                                               To provide estimates of employment, unemployment, and other characteristics of the general labor force, of the population as a whole, and of various subgroups of the population. Monthly labor force data for the country are used by the  Bureau of Labor Statistics (BLS) to determine the distribution of funds under the Job Training Partnership Act. These data are collected through combined computer-assisted personal interviewing (CAPI) and computer-assisted telephone interviewing (CATI). In addition to the labor force data, the CPS basic funding provides annual data on work experience, income, health insurance, and migration data from the Annual Social and Economic Supplement (ASEC), and on school enrollment of the population from the October Supplement. Other supplements, some of which are sponsored by other agencies, are conducted biennially or intermittently.
#> 4                                                                                                                                                                                               To provide estimates of employment, unemployment, and other characteristics of the general labor force, of the population as a whole, and of various subgroups of the population. Monthly labor force data for the country are used by the  Bureau of Labor Statistics (BLS) to determine the distribution of funds under the Job Training Partnership Act. These data are collected through combined computer-assisted personal interviewing (CAPI) and computer-assisted telephone interviewing (CATI). In addition to the labor force data, the CPS basic funding provides annual data on work experience, income, health insurance, and migration data from the Annual Social and Economic Supplement (ASEC), and on school enrollment of the population from the October Supplement. Other supplements, some of which are sponsored by other agencies, are conducted biennially or intermittently.
#> 5                                                                                                                                                                                                To provide estimates of employment, unemployment, and other characteristics of the general labor force, of the population as a whole, and of various subgroups of the population. Monthly labor force data for the country are used by the Bureau of Labor Statistics (BLS) to determine the distribution of funds under the Job Training Partnership Act. These data are collected through combined computer-assisted personal interviewing (CAPI) and computer-assisted telephone interviewing (CATI). In addition to the labor force data, the CPS basic funding provides annual data on work experience, income, health insurance, and migration data from the Annual Social and Economic Supplement (ASEC), and on school enrollment of the population from the October Supplement. Other supplements, some of which are sponsored by other agencies, are conducted biennially or intermittently.
#> 6                                                                                                                                                                                               To provide estimates of employment, unemployment, and other characteristics of the general labor force, of the population as a whole, and of various subgroups of the population. Monthly labor force data for the country are used by the  Bureau of Labor Statistics (BLS) to determine the distribution of funds under the Job Training Partnership Act. These data are collected through combined computer-assisted personal interviewing (CAPI) and computer-assisted telephone interviewing (CATI). In addition to the labor force data, the CPS basic funding provides annual data on work experience, income, health insurance, and migration data from the Annual Social and Economic Supplement (ASEC), and on school enrollment of the population from the October Supplement. Other supplements, some of which are sponsored by other agencies, are conducted biennially or intermittently.
#>              contact
#> 1 dsd.cps@census.gov
#> 2 dsd.cps@census.gov
#> 3 dsd.cps@census.gov
#> 4 dsd.cps@census.gov
#> 5 dsd.cps@census.gov
#> 6 dsd.cps@census.gov

# Get information about all vintage 2022 datasets
apis_2022 <- listCensusApis(vintage = 2022)
head(apis_2022)
#>                                                                     title
#> 1              2022 Annual Business Survey: Characteristics of Businesses
#> 2         2022 Annual Business Survey: Characteristics of Business Owners
#> 3                            2022 Annual Business Survey: Company Summary
#> 4       2022 Annual Business Survey: Module Characteristics of Businesses
#> 5     American Community Survey: 1-Year Estimates: Detailed Tables 1-Year
#> 6 American Community Survey: 1-Year Estimates: Comparison Profiles 1-Year
#>                name vintage      type  temporal       spatial
#> 1             abscb    2022 Aggregate 2022/2022 United States
#> 2            abscbo    2022 Aggregate 2022/2022 United States
#> 3             abscs    2022 Aggregate 2022/2022 United States
#> 4            absmcb    2022 Aggregate 2022/2022 United States
#> 5          acs/acs1    2022 Aggregate 2022/2022            US
#> 6 acs/acs1/cprofile    2022 Aggregate 2022/2022            US
#>                                                 url              modified
#> 1             http://api.census.gov/data/2022/abscb 2024-08-07 17:59:02.0
#> 2            http://api.census.gov/data/2022/abscbo 2024-08-07 17:59:12.0
#> 3             http://api.census.gov/data/2022/abscs 2024-08-07 17:59:09.0
#> 4            http://api.census.gov/data/2022/absmcb 2024-08-07 17:59:06.0
#> 5          http://api.census.gov/data/2022/acs/acs1 2023-04-24 14:51:53.0
#> 6 http://api.census.gov/data/2022/acs/acs1/cprofile 2023-04-24 14:49:08.0
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           description
#> 1                                                                                                                                              The Annual Business Survey (ABS) provides information on selected economic and demographic characteristics for businesses and business owners by sex, ethnicity, race, and veteran status. Further, the survey measures research and development (for microbusinesses), new business topics such as innovation and technology, as well as other business characteristics. The U.S. Census Bureau and the National Center conduct the ABS jointly for Science and Engineering Statistics within the National Science Foundation. The ABS replaces the five-year Survey of Business Owners (SBO) for employer businesses, the Annual Survey of Entrepreneurs (ASE), the Business R&D and Innovation for Microbusinesses survey (BRDI-M), and the innovation section of the Business R&D and Innovation Survey (BRDI-S). https://www.census.gov/programs-surveys/abs.html
#> 2                                                                                                                                              The Annual Business Survey (ABS) provides information on selected economic and demographic characteristics for businesses and business owners by sex, ethnicity, race, and veteran status. Further, the survey measures research and development (for microbusinesses), new business topics such as innovation and technology, as well as other business characteristics. The U.S. Census Bureau and the National Center conduct the ABS jointly for Science and Engineering Statistics within the National Science Foundation. The ABS replaces the five-year Survey of Business Owners (SBO) for employer businesses, the Annual Survey of Entrepreneurs (ASE), the Business R&D and Innovation for Microbusinesses survey (BRDI-M), and the innovation section of the Business R&D and Innovation Survey (BRDI-S). https://www.census.gov/programs-surveys/abs.html
#> 3                                                                                                                                              The Annual Business Survey (ABS) provides information on selected economic and demographic characteristics for businesses and business owners by sex, ethnicity, race, and veteran status. Further, the survey measures research and development (for microbusinesses), new business topics such as innovation and technology, as well as other business characteristics. The U.S. Census Bureau and the National Center conduct the ABS jointly for Science and Engineering Statistics within the National Science Foundation. The ABS replaces the five-year Survey of Business Owners (SBO) for employer businesses, the Annual Survey of Entrepreneurs (ASE), the Business R&D and Innovation for Microbusinesses survey (BRDI-M), and the innovation section of the Business R&D and Innovation Survey (BRDI-S). https://www.census.gov/programs-surveys/abs.html
#> 4                                                                                                                                              The Annual Business Survey (ABS) provides information on selected economic and demographic characteristics for businesses and business owners by sex, ethnicity, race, and veteran status. Further, the survey measures research and development (for microbusinesses), new business topics such as innovation and technology, as well as other business characteristics. The U.S. Census Bureau and the National Center conduct the ABS jointly for Science and Engineering Statistics within the National Science Foundation. The ABS replaces the five-year Survey of Business Owners (SBO) for employer businesses, the Annual Survey of Entrepreneurs (ASE), the Business R&D and Innovation for Microbusinesses survey (BRDI-M), and the innovation section of the Business R&D and Innovation Survey (BRDI-S). https://www.census.gov/programs-surveys/abs.html
#> 5                                                                                               The American Community Survey (ACS) is an ongoing survey that provides data every year -- giving communities the current information they need to plan investments and services. The ACS covers a broad range of topics about social, economic, demographic, and housing characteristics of the U.S. population. Much of the ACS data provided on the Census Bureau's Web site are available separately by age group, race, Hispanic origin, and sex. Summary files, Subject tables, Data profiles, and Comparison profiles are available for the nation, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more. Detailed Tables contain the most detailed cross-tabulations published for areas 65k and more. The data are population counts. There are over 31,000 variables in this dataset.
#> 6 The American Community Survey (ACS) is an ongoing survey that provides data every year -- giving communities the current information they need to plan investments and services. The ACS covers a broad range of topics about social, economic, demographic, and housing characteristics of the U.S. population. Much of the ACS data provided on the Census Bureau's Web site are available separately by age group, race, Hispanic origin, and sex. Summary files, Subject tables, Data profiles, and Comparison profiles are available for the nation, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more. Comparison profiles are similar to Data profiles but also include comparisons with past-year data. The current year data are compared with each of the last four years of data and include statistical significance testing. There are over 1,000 variables in this dataset.
#>                                  contact
#> 1 adep.annual.business.survey@census.gov
#> 2 adep.annual.business.survey@census.gov
#> 3 adep.annual.business.survey@census.gov
#> 4 adep.annual.business.survey@census.gov
#> 5          acso.users.support@census.gov
#> 6          acso.users.support@census.gov

# Get information about all timeseries datasets
apis_timeseries <- listCensusApis(name = "timeseries")
head(apis_timeseries)
#>                                                                                      title
#> 1                              Annual Economic Surveys: Annual Survey of Manufactures Area
#> 2      Economic Surveys: Annual Survey of Manufactures: Annual Survey of Manufactures Area
#> 3                    Annual Economic Surveys: Annual Survey of Manufactures Benchmark 2017
#> 4 Time Series Annual Survey of Manufactures: Statistics for Industry Groups and Industries
#> 5        Time Series Annual Survey of Manufactures: Value of Shipments for Product Classes
#> 6     Time Series Annual Survey of Manufactures: Statistics for All Manufacturing by State
#>                           name       type  temporal       spatial
#> 1      timeseries/asm/area2012 Timeseries      <NA>          <NA>
#> 2      timeseries/asm/area2017 Timeseries 2018/2021 United States
#> 3 timeseries/asm/benchmark2017 Timeseries      <NA>          <NA>
#> 4      timeseries/asm/industry Timeseries      <NA> United States
#> 5       timeseries/asm/product Timeseries      <NA> United States
#> 6         timeseries/asm/state Timeseries      <NA> United States
#>                                                       url              modified
#> 1      http://api.census.gov/data/timeseries/asm/area2012 2018-12-13 00:00:00.0
#> 2      http://api.census.gov/data/timeseries/asm/area2017 2020-03-17 00:00:00.0
#> 3 http://api.census.gov/data/timeseries/asm/benchmark2017 2021-03-17 00:00:00.0
#> 4      http://api.census.gov/data/timeseries/asm/industry            2018-06-29
#> 5       http://api.census.gov/data/timeseries/asm/product            2017-12-15
#> 6         http://api.census.gov/data/timeseries/asm/state            2017-12-15
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   description
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The Annual Survey of Manufactures (ASM) provides key intercensal measures of manufacturing activity, products, and location for the public and private sectors. The ASM provides the best current measure of current U.S. manufacturing industry outputs, inputs, and operating status, and is the primary basis for updates of the Longitudinal Research Database (LRD). Census Bureau staff and academic researchers with sworn agent status use the LRD for micro data analysis.
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The Annual Survey of Manufactures (ASM) provides key intercensal measures of manufacturing activity, products, and location for the public and private sectors. The ASM provides the best current measure of current U.S. manufacturing industry outputs, inputs, and operating status, and is the primary basis for updates of the Longitudinal Research Database (LRD). Census Bureau staff and academic researchers with sworn agent status use the LRD for micro data analysis.
#> 3 The Annual Survey of Manufactures (ASM) Benchmark provides key intercensal measures of manufacturing activity and products for the public and private sectors for four years following the Economic Census of Manufacturing. These benchmark tables present manufacturing establishment statistics from the 2013-2016 Annual Survey of Manufactures (ASM). The ASM Benchmark provides statistics on employment, payroll, worker hours, cost of materials, value added by manufacturing, inventories, and estimates for value of shipments for product classes of products manufactured as defined by the North American Industry Classification System. The ASM provides the best current measure of current U.S. manufacturing industry outputs, inputs, and operating status, and is the primary basis for updates of the Longitudinal Research Database (LRD). Census Bureau staff and academic researchers with sworn agent status use the LRD for micro data analysis.
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The Annual Survey of Manufactures (ASM) provides key intercensal measures of manufacturing activity, products, and location for the public and private sectors. The ASM provides the best current measure of current U.S. manufacturing industry outputs, inputs, and operating status, and is the primary basis for updates of the Longitudinal Research Database (LRD). Census Bureau staff and academic researchers with sworn agent status use the LRD for micro data analysis.
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The Annual Survey of Manufactures (ASM) provides key intercensal measures of manufacturing activity, products, and location for the public and private sectors. The ASM provides the best current measure of current U.S. manufacturing industry outputs, inputs, and operating status, and is the primary basis for updates of the Longitudinal Research Database (LRD). Census Bureau staff and academic researchers with sworn agent status use the LRD for micro data analysis.
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The Annual Survey of Manufactures (ASM) provides key intercensal measures of manufacturing activity, products, and location for the public and private sectors. The ASM provides the best current measure of current U.S. manufacturing industry outputs, inputs, and operating status, and is the primary basis for updates of the Longitudinal Research Database (LRD). Census Bureau staff and academic researchers with sworn agent status use the LRD for micro data analysis.
#>                         contact
#> 1       Ewd.outreach@census.gov
#> 2       ewd.outreach@census.gov
#> 3       Ewd.outreach@census.gov
#> 4 econ.dissemination@census.gov
#> 5   stephen.c.mangum@census.gov
#> 6   stephen.c.mangum@census.gov

# Get information about 2020 Decennial Census datasets
apis_decennial_2020 <- listCensusApis(name = "dec", vintage = 2020)
head(apis_decennial_2020)
#>                                                                                                      title
#> 1                                              Decennial Census: 118th Congressional District Summary File
#> 2                                                                 2020 Census 119th Congressional District
#> 3                               Decennial Census of Island Areas: American Samoa Detailed Crosstabulations
#> 4                                         Decennial Census of Island Areas: Guam Detailed Crosstabulations
#> 5 Decennial Census of Island Areas: Commonwealth of the Northern Mariana Islands Detailed Crosstabulations
#> 6                          Decennial Census of Island Areas: U.S. Virgin Islands Detailed Crosstabulations
#>             name vintage      type  temporal                  spatial
#> 1      dec/cd118    2020 Aggregate 2020/2020                       US
#> 2      dec/cd119    2020 Aggregate 2020/2020                       US
#> 3 dec/crosstabas    2020 Aggregate 2020/2020           American Samoa
#> 4 dec/crosstabgu    2020 Aggregate 2020/2020                     Guam
#> 5 dec/crosstabmp    2020 Aggregate 2020/2020 Northern Mariana Islands
#> 6 dec/crosstabvi    2020 Aggregate 2020/2020      U.S. Virgin Islands
#>                                              url              modified
#> 1      http://api.census.gov/data/2020/dec/cd118 2022-10-25 00:00:00.0
#> 2      http://api.census.gov/data/2020/dec/cd119 2024-02-15 10:18:36.0
#> 3 http://api.census.gov/data/2020/dec/crosstabas 2023-05-09 10:44:24.0
#> 4 http://api.census.gov/data/2020/dec/crosstabgu 2023-05-09 10:46:02.0
#> 5 http://api.census.gov/data/2020/dec/crosstabmp 2023-05-09 10:45:29.0
#> 6 http://api.census.gov/data/2020/dec/crosstabvi 2023-05-09 10:44:57.0
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  description
#> 1 The Congressional District Summary File (118th Congress) (CD118) contains the data compiled from the questions asked of all people and about every housing unit in the 2020 Census. This product retabulates selected summary levels from the Demographic and Housing Characteristics File (DHC) for the 118th Congress and 2022 state legislative districts. Population items include age, sex, race, Hispanic or Latino origin, household type, family type, relationship to householder, group quarters population, housing occupancy and housing tenure (whether a housing unit is owner-occupied or renter-occupied).
#> 2 The Congressional District Summary File (119th Congress) (CD119) contains the data compiled from the questions asked of all people and about every housing unit in the 2020 Census. This product retabulates selected summary levels from the Demographic and Housing Characteristics File (DHC) for the 119th Congress and 2024 state legislative districts. Population items include age, sex, race, Hispanic or Latino origin, household type, family type, relationship to householder, group quarters population, housing occupancy and housing tenure (whether a housing unit is owner-occupied or renter-occupied).
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                              This product will include key socio-demographic and economic variables presented in cross-tabulations that present the complex data together.
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                This product will include….
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                              This product will include key socio-demographic and economic variables presented in cross-tabulations that present the complex data together.
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                              This product will include key socio-demographic and economic variables presented in cross-tabulations that present the complex data together.
#>          contact
#> 1 pio@census.gov
#> 2 pio@census.gov
#> 3 pio@census.gov
#> 4 pio@census.gov
#> 5 pio@census.gov
#> 6 pio@census.gov

# Get information about one particular dataset
api_sahie <- listCensusApis(name = "timeseries/healthins/sahie")
head(api_sahie)
#>                                               title                       name
#> 1 Time Series Small Area Health Insurance Estimates timeseries/healthins/sahie
#>         type  temporal spatial
#> 1 Timeseries 2006/2022      US
#>                                                     url              modified
#> 1 http://api.census.gov/data/timeseries/healthins/sahie 2021-04-12 00:00:00.0
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         description
#> 1 The U.S. Census Bureau's Small Area Health Insurance Estimates program produces the only source of data for single-year estimates of health insurance coverage status for all counties in the U.S. by selected economic and demographic characteristics. This program is partially funded by the Centers for Disease Control and Prevention's (CDC) Division of Cancer Prevention and Control (DCPC). The CDC have a congressional mandate to provide screening services for breast and cervical cancer to low-income, uninsured, and underserved women through the National Breast and Cervical Cancer Early Detection Program (NBCCEDP). For estimation, SAHIE uses statistical models that combine survey data from the American Community Survey (ACS) with administrative records data and Census 2020 data.
#>                  contact
#> 1 Sehsd.sahie@census.gov