Get dataset metadata on all available APIs as a data frame

Scrapes https://api.census.gov/data.json and returns a dataframe that includes: title, name, vintage (where applicable), url, isTimeseries (binary), temporal (helpful for some time series), description, modified date

listCensusApis()

Examples

apis <- listCensusApis()
head(apis)

#>                                                    title      name vintage
#> 23  2005-2009 American Community Survey: Migration Flows acs/flows    2009
#> 136 2006-2010 American Community Survey: Migration Flows acs/flows    2010
#> 73  2007-2011 American Community Survey: Migration Flows acs/flows    2011
#> 144 2008-2012 American Community Survey: Migration Flows acs/flows    2012
#> 160 2009-2013 American Community Survey: Migration Flows acs/flows    2013
#> 70  2010-2014 American Community Survey: Migration Flows acs/flows    2014
#>                                            url isTimeseries  temporal
#> 23  https://api.census.gov/data/2009/acs/flows           NA 2006-2010
#> 136 https://api.census.gov/data/2010/acs/flows           NA 2006-2010
#> 73  https://api.census.gov/data/2011/acs/flows           NA 2006-2010
#> 144 https://api.census.gov/data/2012/acs/flows           NA 2006-2010
#> 160 https://api.census.gov/data/2013/acs/flows           NA 2006-2010
#> 70  https://api.census.gov/data/2014/acs/flows           NA 2006-2010
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        description
#> 23  Migration flows are derived from the relationship between the location of current residence in the American Community Survey (ACS) sample and the responses given to the migration question "Where did you live 1 year ago?".  There are flow statistics (moved in, moved out, and net moved) between county or minor civil division (MCD) of residence and county, MCD, or world region of residence 1 year ago.  Estimates for MCDs are only available for the 12 strong-MCD states, where the MCDs have the same government functions as incorporated places.  Migration flows between metropolitan statistical areas are available starting with the 2009-2013 5-year ACS dataset.   Flow statistics are available by three or four variables for each dataset starting with the 2006-2010 5-year ACS datasets.  The variables change for each dataset and do not repeat in overlapping datasets.  In addition to the flow estimates, there are supplemental statistics files that contain migration/geographical mobility estimates (e.g., nonmovers, moved to a different state, moved from abroad) for each county, MCD, or metro area.
#> 136 Migration flows are derived from the relationship between the location of current residence in the American Community Survey (ACS) sample and the responses given to the migration question "Where did you live 1 year ago?".  There are flow statistics (moved in, moved out, and net moved) between county or minor civil division (MCD) of residence and county, MCD, or world region of residence 1 year ago.  Estimates for MCDs are only available for the 12 strong-MCD states, where the MCDs have the same government functions as incorporated places.  Migration flows between metropolitan statistical areas are available starting with the 2009-2013 5-year ACS dataset.   Flow statistics are available by three or four variables for each dataset starting with the 2006-2010 5-year ACS datasets.  The variables change for each dataset and do not repeat in overlapping datasets.  In addition to the flow estimates, there are supplemental statistics files that contain migration/geographical mobility estimates (e.g., nonmovers, moved to a different state, moved from abroad) for each county, MCD, or metro area.
#> 73  Migration flows are derived from the relationship between the location of current residence in the American Community Survey (ACS) sample and the responses given to the migration question "Where did you live 1 year ago?".  There are flow statistics (moved in, moved out, and net moved) between county or minor civil division (MCD) of residence and county, MCD, or world region of residence 1 year ago.  Estimates for MCDs are only available for the 12 strong-MCD states, where the MCDs have the same government functions as incorporated places.  Migration flows between metropolitan statistical areas are available starting with the 2009-2013 5-year ACS dataset.   Flow statistics are available by three or four variables for each dataset starting with the 2006-2010 5-year ACS datasets.  The variables change for each dataset and do not repeat in overlapping datasets.  In addition to the flow estimates, there are supplemental statistics files that contain migration/geographical mobility estimates (e.g., nonmovers, moved to a different state, moved from abroad) for each county, MCD, or metro area.
#> 144 Migration flows are derived from the relationship between the location of current residence in the American Community Survey (ACS) sample and the responses given to the migration question "Where did you live 1 year ago?".  There are flow statistics (moved in, moved out, and net moved) between county or minor civil division (MCD) of residence and county, MCD, or world region of residence 1 year ago.  Estimates for MCDs are only available for the 12 strong-MCD states, where the MCDs have the same government functions as incorporated places.  Migration flows between metropolitan statistical areas are available starting with the 2009-2013 5-year ACS dataset.   Flow statistics are available by three or four variables for each dataset starting with the 2006-2010 5-year ACS datasets.  The variables change for each dataset and do not repeat in overlapping datasets.  In addition to the flow estimates, there are supplemental statistics files that contain migration/geographical mobility estimates (e.g., nonmovers, moved to a different state, moved from abroad) for each county, MCD, or metro area.
#> 160 Migration flows are derived from the relationship between the location of current residence in the American Community Survey (ACS) sample and the responses given to the migration question "Where did you live 1 year ago?".  There are flow statistics (moved in, moved out, and net moved) between county or minor civil division (MCD) of residence and county, MCD, or world region of residence 1 year ago.  Estimates for MCDs are only available for the 12 strong-MCD states, where the MCDs have the same government functions as incorporated places.  Migration flows between metropolitan statistical areas are available starting with the 2009-2013 5-year ACS dataset.   Flow statistics are available by three or four variables for each dataset starting with the 2006-2010 5-year ACS datasets.  The variables change for each dataset and do not repeat in overlapping datasets.  In addition to the flow estimates, there are supplemental statistics files that contain migration/geographical mobility estimates (e.g., nonmovers, moved to a different state, moved from abroad) for each county, MCD, or metro area.
#> 70  Migration flows are derived from the relationship between the location of current residence in the American Community Survey (ACS) sample and the responses given to the migration question "Where did you live 1 year ago?".  There are flow statistics (moved in, moved out, and net moved) between county or minor civil division (MCD) of residence and county, MCD, or world region of residence 1 year ago.  Estimates for MCDs are only available for the 12 strong-MCD states, where the MCDs have the same government functions as incorporated places.  Migration flows between metropolitan statistical areas are available starting with the 2009-2013 5-year ACS dataset.   Flow statistics are available by three or four variables for each dataset starting with the 2006-2010 5-year ACS datasets.  The variables change for each dataset and do not repeat in overlapping datasets.  In addition to the flow estimates, there are supplemental statistics files that contain migration/geographical mobility estimates (e.g., nonmovers, moved to a different state, moved from abroad) for each county, MCD, or metro area.
#>       modified
#> 23  2016-12-23
#> 136 2016-06-22
#> 73  2016-06-22
#> 144 2016-06-22
#> 160 2016-06-22
#> 70  2016-11-15