Get dataset metadata on all available APIs as a data frame
Source: R/metadata_functions.R

Scrapes https://api.census.gov/data.json and returns a dataframe that includes: title, name, vintage (where applicable), url, isTimeseries (binary), temporal (helpful for some time series), description, modified date

listCensusApis()

Examples

apis <- listCensusApis()
head(apis)

#>                             title              name vintage
#> 104    ACS 1-Year Detailed Tables          acs/acs1    2015
#> 119    ACS 1-Year Detailed Tables          acs/acs1    2016
#> 168 ACS 1-Year Comparison Profile acs/acs1/cprofile    2015
#> 129 ACS 1-Year Comparison Profile acs/acs1/cprofile    2016
#> 78       ACS 1-Year Data Profiles  acs/acs1/profile    2015
#> 36       ACS 1-Year Data Profiles  acs/acs1/profile    2016
#>                                                    url isTimeseries
#> 104          https://api.census.gov/data/2015/acs/acs1           NA
#> 119          https://api.census.gov/data/2016/acs/acs1           NA
#> 168 https://api.census.gov/data/2015/acs/acs1/cprofile           NA
#> 129 https://api.census.gov/data/2016/acs/acs1/cprofile           NA
#> 78   https://api.census.gov/data/2015/acs/acs1/profile           NA
#> 36   https://api.census.gov/data/2016/acs/acs1/profile           NA
#>         temporal
#> 104 unidentified
#> 119 unidentified
#> 168 unidentified
#> 129 unidentified
#> 78  unidentified
#> 36  unidentified
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        description
#> 104                 The American Community Survey (ACS) is a uswide survey designed to provide communities a fresh look at how they are changing. The ACS replaced the decennial census long form in 2010 and thereafter by collecting long form type information throughout the decade rather than only once every 10 years. Questionnaires are mailed to a sample of addresses to obtain information about households -- that is, about each person and the housing unit itself. The American Community Survey produces demographic, social, housing and economic estimates in the form of 1 and 5-year estimates based on population thresholds. The strength of the ACS is in estimating population and housing characteristics. The 2012 data provide key estimates for each of the topic areas covered by the ACS for the us, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more. Although the ACS produces population, demographic and housing unit estimates,it is the Census Bureau's Population Estimates Program that produces and disseminates the official estimates of the population for the us, states, counties, cities and towns, and estimates of housing units for states and counties. For 2010 and other decennial census years, the Decennial Census provides the official counts of population and housing units.
#> 119                 The American Community Survey (ACS) is a uswide survey designed to provide communities a fresh look at how they are changing. The ACS replaced the decennial census long form in 2010 and thereafter by collecting long form type information throughout the decade rather than only once every 10 years. Questionnaires are mailed to a sample of addresses to obtain information about households -- that is, about each person and the housing unit itself. The American Community Survey produces demographic, social, housing and economic estimates in the form of 1 and 5-year estimates based on population thresholds. The strength of the ACS is in estimating population and housing characteristics. The 2012 data provide key estimates for each of the topic areas covered by the ACS for the us, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more. Although the ACS produces population, demographic and housing unit estimates,it is the Census Bureau's Population Estimates Program that produces and disseminates the official estimates of the population for the us, states, counties, cities and towns, and estimates of housing units for states and counties. For 2010 and other decennial census years, the Decennial Census provides the official counts of population and housing units.
#> 168                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  The American Community Survey (ACS) is an ongoing survey that provides data every year -- giving communities the current information they need to plan investments and services. The ACS covers a broad range of topics about social, economic, demographic, and housing characteristics of the U.S. population.  Summary file, Subject tables, Data profiles, and Comparison profiles are available for the us, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more.  Comparison profiles are similar to data profiles but also include comparisons with past-year data.  The current year data are compared with each of the last four years of data and include statistical significance testing.  There are over 9,000 variables in this dataset.
#> 129                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  The American Community Survey (ACS) is an ongoing survey that provides data every year -- giving communities the current information they need to plan investments and services. The ACS covers a broad range of topics about social, economic, demographic, and housing characteristics of the U.S. population.  Summary file, Subject tables, Data profiles, and Comparison profiles are available for the us, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more.  Comparison profiles are similar to data profiles but also include comparisons with past-year data.  The current year data are compared with each of the last four years of data and include statistical significance testing.  There are over 9,000 variables in this dataset.
#> 78  The American Community Survey (ACS) is a nationwide survey designed to provide communities a fresh look at how they are changing. The ACS replaced the decennial census long form in 2010 and thereafter by collecting long form type information throughout the decade rather than only once every 10 years. Questionnaires are mailed to a sample of addresses to obtain information about households -- that is, about each person and the housing unit itself. The American Community Survey produces demographic, social, housing and economic estimates in the form of 1 and 5-year estimates based on population thresholds. The strength of the ACS is in estimating population and housing characteristics. The data profiles provide key estimates for each of the topic areas covered by the ACS for the nation, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more. Although the ACS produces population, demographic and housing unit estimates,it is the Census Bureau's Population Estimates Program that produces and disseminates the official estimates of the population for the nation, states, counties, cities and towns, and estimates of housing units for states and counties. For 2010 and other decennial census years, the Decennial Census provides the official counts of population and housing units.
#> 36              The American Community Survey (ACS) is a uswide survey designed to provide communities a fresh look at how they are changing. The ACS replaced the decennial census long form in 2010 and thereafter by collecting long form type information throughout the decade rather than only once every 10 years. Questionnaires are mailed to a sample of addresses to obtain information about households -- that is, about each person and the housing unit itself. The American Community Survey produces demographic, social, housing and economic estimates in the form of 1 and 5-year estimates based on population thresholds. The strength of the ACS is in estimating population and housing characteristics. The data profiles provide key estimates for each of the topic areas covered by the ACS for the us, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more. Although the ACS produces population, demographic and housing unit estimates,it is the Census Bureau's Population Estimates Program that produces and disseminates the official estimates of the population for the us, states, counties, cities and towns, and estimates of housing units for states and counties. For 2010 and other decennial census years, the Decennial Census provides the official counts of population and housing units.
#>       modified
#> 104 2017-12-14
#> 119 2017-12-14
#> 168 2017-12-14
#> 129 2017-12-14
#> 78  2017-12-14
#> 36  2017-12-14