Get dataset metadata on all available APIs as a data frame

Scrapes https://api.census.gov/data.json and returns a dataframe that includes: title, name, vintage (where applicable), url, isTimeseries (binary), temporal (helpful for some time series), description, modified date

listCensusApis()

Examples

apis <- listCensusApis()
head(apis)

#>                                                                   title
#> 396 American Community Survey: 1-Year Estimates: Detailed Tables 1-Year
#> 400 American Community Survey: 1-Year Estimates: Detailed Tables 1-Year
#> 402 American Community Survey: 1-Year Estimates: Detailed Tables 1-Year
#> 404 American Community Survey: 1-Year Estimates: Detailed Tables 1-Year
#> 408 American Community Survey: 1-Year Estimates: Detailed Tables 1-Year
#> 75                                           ACS 1-Year Detailed Tables
#>         name vintage                                       url isTimeseries
#> 396 acs/acs1    2005 https://api.census.gov/data/2005/acs/acs1           NA
#> 400 acs/acs1    2006 https://api.census.gov/data/2006/acs/acs1           NA
#> 402 acs/acs1    2007 https://api.census.gov/data/2007/acs/acs1           NA
#> 404 acs/acs1    2008 https://api.census.gov/data/2008/acs/acs1           NA
#> 408 acs/acs1    2009 https://api.census.gov/data/2009/acs/acs1           NA
#> 75  acs/acs1    2010 https://api.census.gov/data/2010/acs/acs1           NA
#>         temporal
#> 396 unidentified
#> 400 unidentified
#> 402 unidentified
#> 404 unidentified
#> 408 unidentified
#> 75  unidentified
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    description
#> 396                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      The American Community Survey (ACS) is an ongoing survey that provides data every year -- giving communities the current information they need to plan investments and services. The ACS covers a broad range of topics about social, economic, demographic, and housing characteristics of the U.S. population. Much of the ACS data provided on the Census Bureau's Web site are available separately by age group, race, Hispanic origin, and sex.  Summary files, Subject tables, Data profiles, and Comparison profiles are available for the nation, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more.  Detail Tables contain the most detailed cross-tabulations published for areas 65k and more. The data are population counts. There are over 31,000 variables in this dataset.
#> 400                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      The American Community Survey (ACS) is an ongoing survey that provides data every year -- giving communities the current information they need to plan investments and services. The ACS covers a broad range of topics about social, economic, demographic, and housing characteristics of the U.S. population. Much of the ACS data provided on the Census Bureau's Web site are available separately by age group, race, Hispanic origin, and sex.  Summary files, Subject tables, Data profiles, and Comparison profiles are available for the nation, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more.  Detail Tables contain the most detailed cross-tabulations published for areas 65k and more. The data are population counts. There are over 31,000 variables in this dataset.
#> 402                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      The American Community Survey (ACS) is an ongoing survey that provides data every year -- giving communities the current information they need to plan investments and services. The ACS covers a broad range of topics about social, economic, demographic, and housing characteristics of the U.S. population. Much of the ACS data provided on the Census Bureau's Web site are available separately by age group, race, Hispanic origin, and sex.  Summary files, Subject tables, Data profiles, and Comparison profiles are available for the nation, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more.  Detail Tables contain the most detailed cross-tabulations published for areas 65k and more. The data are population counts. There are over 31,000 variables in this dataset.
#> 404                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      The American Community Survey (ACS) is an ongoing survey that provides data every year -- giving communities the current information they need to plan investments and services. The ACS covers a broad range of topics about social, economic, demographic, and housing characteristics of the U.S. population. Much of the ACS data provided on the Census Bureau's Web site are available separately by age group, race, Hispanic origin, and sex.  Summary files, Subject tables, Data profiles, and Comparison profiles are available for the nation, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more.  Detail Tables contain the most detailed cross-tabulations published for areas 65k and more. The data are population counts. There are over 31,000 variables in this dataset.
#> 408                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      The American Community Survey (ACS) is an ongoing survey that provides data every year -- giving communities the current information they need to plan investments and services. The ACS covers a broad range of topics about social, economic, demographic, and housing characteristics of the U.S. population. Much of the ACS data provided on the Census Bureau's Web site are available separately by age group, race, Hispanic origin, and sex.  Summary files, Subject tables, Data profiles, and Comparison profiles are available for the nation, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more.  Detail Tables contain the most detailed cross-tabulations published for areas 65k and more. The data are population counts. There are over 31,000 variables in this dataset.
#> 75  The American Community Survey (ACS) is a nationwide survey designed to provide communities a fresh look at how they are changing. The ACS replaced the decennial census long form in 2010 and thereafter by collecting long form type information throughout the decade rather than only once every 10 years. Questionnaires are mailed to a sample of addresses to obtain information about households -- that is, about each person and the housing unit itself. The American Community Survey produces demographic, social, housing and economic estimates in the form of 1 and 5-year estimates based on population thresholds. The strength of the ACS is in estimating population and housing characteristics. The 2012 data provide key estimates for each of the topic areas covered by the ACS for the nation, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more. Although the ACS produces population, demographic and housing unit estimates,it is the Census Bureau's Population Estimates Program that produces and disseminates the official estimates of the population for the nation, states, counties, cities and towns, and estimates of housing units for states and counties. For 2010 and other decennial census years, the Decennial Census provides the official counts of population and housing units.
#>                  modified
#> 396 2019-08-27 12:50:06.0
#> 400 2019-08-27 10:47:47.0
#> 402 2019-08-27 10:05:20.0
#> 404 2019-08-27 07:30:33.0
#> 408 2019-08-26 10:52:39.0
#> 75  2018-07-04 00:00:00.0