Get dataset metadata on all available APIs as a data frame

Scrapes https://api.census.gov/data.json and returns a dataframe that includes: title, name, vintage (where applicable), url, isTimeseries (binary), temporal (helpful for some time series), description, modified date

listCensusApis()

Examples

apis <- listCensusApis()
head(apis)

#>                                                                   title
#> 456                     Annual Economic Surveys: Annual Business Survey
#> 460                     Annual Economic Surveys: Annual Business Survey
#> 462                     Annual Economic Surveys: Annual Business Survey
#> 397 American Community Survey: 1-Year Estimates: Detailed Tables 1-Year
#> 401 American Community Survey: 1-Year Estimates: Detailed Tables 1-Year
#> 403 American Community Survey: 1-Year Estimates: Detailed Tables 1-Year
#>         name vintage                                       url isTimeseries
#> 456    abscb    2017    https://api.census.gov/data/2017/abscb           NA
#> 460   abscbo    2017   https://api.census.gov/data/2017/abscbo           NA
#> 462    abscs    2017    https://api.census.gov/data/2017/abscs           NA
#> 397 acs/acs1    2005 https://api.census.gov/data/2005/acs/acs1           NA
#> 401 acs/acs1    2006 https://api.census.gov/data/2006/acs/acs1           NA
#> 403 acs/acs1    2007 https://api.census.gov/data/2007/acs/acs1           NA
#>         temporal
#> 456 unidentified
#> 460 unidentified
#> 462 unidentified
#> 397 unidentified
#> 401 unidentified
#> 403 unidentified
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               description
#> 456                                                The Annual Business Survey (ABS) provides information on selected economic and demographic characteristics for businesses and business owners by sex, ethnicity, race, and veteran status. Further, the survey measures research and development (for microbusinesses), new business topics such as innovation and technology, as well as other business characteristics. The U.S. Census Bureau and the National Center conduct the ABS jointly for Science and Engineering Statistics within the National Science Foundation. The ABS replaces the five-year Survey of Business Owners (SBO) for employer businesses, the Annual Survey of Entrepreneurs (ASE), the Business R&D and Innovation for Microbusinesses survey (BRDI-M), and the innovation section of the Business R&D and Innovation Survey (BRDI-S). https://www.census.gov/programs-surveys/abs.html
#> 460                                                The Annual Business Survey (ABS) provides information on selected economic and demographic characteristics for businesses and business owners by sex, ethnicity, race, and veteran status. Further, the survey measures research and development (for microbusinesses), new business topics such as innovation and technology, as well as other business characteristics. The U.S. Census Bureau and the National Center conduct the ABS jointly for Science and Engineering Statistics within the National Science Foundation. The ABS replaces the five-year Survey of Business Owners (SBO) for employer businesses, the Annual Survey of Entrepreneurs (ASE), the Business R&D and Innovation for Microbusinesses survey (BRDI-M), and the innovation section of the Business R&D and Innovation Survey (BRDI-S). https://www.census.gov/programs-surveys/abs.html
#> 462                                                The Annual Business Survey (ABS) provides information on selected economic and demographic characteristics for businesses and business owners by sex, ethnicity, race, and veteran status. Further, the survey measures research and development (for microbusinesses), new business topics such as innovation and technology, as well as other business characteristics. The U.S. Census Bureau and the National Center conduct the ABS jointly for Science and Engineering Statistics within the National Science Foundation. The ABS replaces the five-year Survey of Business Owners (SBO) for employer businesses, the Annual Survey of Entrepreneurs (ASE), the Business R&D and Innovation for Microbusinesses survey (BRDI-M), and the innovation section of the Business R&D and Innovation Survey (BRDI-S). https://www.census.gov/programs-surveys/abs.html
#> 397 The American Community Survey (ACS) is an ongoing survey that provides data every year -- giving communities the current information they need to plan investments and services. The ACS covers a broad range of topics about social, economic, demographic, and housing characteristics of the U.S. population. Much of the ACS data provided on the Census Bureau's Web site are available separately by age group, race, Hispanic origin, and sex.  Summary files, Subject tables, Data profiles, and Comparison profiles are available for the nation, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more.  Detail Tables contain the most detailed cross-tabulations published for areas 65k and more. The data are population counts. There are over 31,000 variables in this dataset.
#> 401 The American Community Survey (ACS) is an ongoing survey that provides data every year -- giving communities the current information they need to plan investments and services. The ACS covers a broad range of topics about social, economic, demographic, and housing characteristics of the U.S. population. Much of the ACS data provided on the Census Bureau's Web site are available separately by age group, race, Hispanic origin, and sex.  Summary files, Subject tables, Data profiles, and Comparison profiles are available for the nation, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more.  Detail Tables contain the most detailed cross-tabulations published for areas 65k and more. The data are population counts. There are over 31,000 variables in this dataset.
#> 403 The American Community Survey (ACS) is an ongoing survey that provides data every year -- giving communities the current information they need to plan investments and services. The ACS covers a broad range of topics about social, economic, demographic, and housing characteristics of the U.S. population. Much of the ACS data provided on the Census Bureau's Web site are available separately by age group, race, Hispanic origin, and sex.  Summary files, Subject tables, Data profiles, and Comparison profiles are available for the nation, all 50 states, the District of Columbia, Puerto Rico, every congressional district, every metropolitan area, and all counties and places with populations of 65,000 or more.  Detail Tables contain the most detailed cross-tabulations published for areas 65k and more. The data are population counts. There are over 31,000 variables in this dataset.
#>                  modified
#> 456 2020-04-30 00:00:00.0
#> 460 2020-04-30 00:00:00.0
#> 462 2020-04-30 00:00:00.0
#> 397 2019-08-27 12:50:06.0
#> 401 2019-08-27 10:47:47.0
#> 403 2019-08-27 10:05:20.0