Get useful dataset metadata on all available APIs as a data frame

Scrapes https://api.census.gov/data.json and returns a dataframe that includes: title, description, name, vintage, url, dataset type, and other useful fields.

Usage

listCensusApis()

Examples

apis <- listCensusApis()
head(apis)
#>                                      title          name vintage      type
#> 1 Current Population Survey: Basic Monthly cps/basic/apr    2022 Microdata
#> 2 Current Population Survey: Basic Monthly cps/basic/feb    2022 Microdata
#> 3 Current Population Survey: Basic Monthly cps/basic/jan    2022 Microdata
#> 4 Current Population Survey: Basic Monthly cps/basic/jun    2022 Microdata
#> 5 Current Population Survey: Basic Monthly cps/basic/mar    2022 Microdata
#> 6 Current Population Survey: Basic Monthly cps/basic/may    2022 Microdata
#>          temporal                                           url
#> 1 2022-04/2022-04 http://api.census.gov/data/2022/cps/basic/apr
#> 2 2022-02/2022-02 http://api.census.gov/data/2022/cps/basic/feb
#> 3 2022-01/2022-01 http://api.census.gov/data/2022/cps/basic/jan
#> 4 2022-06/2022-06 http://api.census.gov/data/2022/cps/basic/jun
#> 5 2022-03/2022-03 http://api.census.gov/data/2022/cps/basic/mar
#> 6 2022-05/2022-05 http://api.census.gov/data/2022/cps/basic/may
#>                modified
#> 1 2022-01-19 00:00:00.0
#> 2 2022-01-19 00:00:00.0
#> 3 2022-01-19 00:00:00.0
#> 4 2022-01-19 00:00:00.0
#> 5 2022-01-19 00:00:00.0
#> 6 2022-01-19 00:00:00.0
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          description
#> 1 To provide estimates of employment, unemployment, and other characteristics of the general labor force, of the population as a whole, and of various subgroups of the population. Monthly labor force data for the country are used by the &lt;ahref="http://www.bls.gov"&gt;Bureau of Labor Statistics (BLS)&lt;/a&gt; to determine the distribution of funds under the Job Training Partnership Act. These data are collected through combined computer-assisted personal interviewing (CAPI) and computer-assisted telephone interviewing (CATI). In addition to the labor force data, the CPS basic funding provides annual data on work experience, income, health insurance, and migration data from the Annual Social and Economic Supplement (ASEC), and on school enrollment of the population from the October Supplement. Other supplements, some of which are sponsored by other agencies, are conducted biennially or intermittently.
#> 2 To provide estimates of employment, unemployment, and other characteristics of the general labor force, of the population as a whole, and of various subgroups of the population. Monthly labor force data for the country are used by the &lt;ahref="http://www.bls.gov"&gt;Bureau of Labor Statistics (BLS)&lt;/a&gt; to determine the distribution of funds under the Job Training Partnership Act. These data are collected through combined computer-assisted personal interviewing (CAPI) and computer-assisted telephone interviewing (CATI). In addition to the labor force data, the CPS basic funding provides annual data on work experience, income, health insurance, and migration data from the Annual Social and Economic Supplement (ASEC), and on school enrollment of the population from the October Supplement. Other supplements, some of which are sponsored by other agencies, are conducted biennially or intermittently.
#> 3 To provide estimates of employment, unemployment, and other characteristics of the general labor force, of the population as a whole, and of various subgroups of the population. Monthly labor force data for the country are used by the &lt;ahref="http://www.bls.gov"&gt;Bureau of Labor Statistics (BLS)&lt;/a&gt; to determine the distribution of funds under the Job Training Partnership Act. These data are collected through combined computer-assisted personal interviewing (CAPI) and computer-assisted telephone interviewing (CATI). In addition to the labor force data, the CPS basic funding provides annual data on work experience, income, health insurance, and migration data from the Annual Social and Economic Supplement (ASEC), and on school enrollment of the population from the October Supplement. Other supplements, some of which are sponsored by other agencies, are conducted biennially or intermittently.
#> 4 To provide estimates of employment, unemployment, and other characteristics of the general labor force, of the population as a whole, and of various subgroups of the population. Monthly labor force data for the country are used by the &lt;ahref="http://www.bls.gov"&gt;Bureau of Labor Statistics (BLS)&lt;/a&gt; to determine the distribution of funds under the Job Training Partnership Act. These data are collected through combined computer-assisted personal interviewing (CAPI) and computer-assisted telephone interviewing (CATI). In addition to the labor force data, the CPS basic funding provides annual data on work experience, income, health insurance, and migration data from the Annual Social and Economic Supplement (ASEC), and on school enrollment of the population from the October Supplement. Other supplements, some of which are sponsored by other agencies, are conducted biennially or intermittently.
#> 5 To provide estimates of employment, unemployment, and other characteristics of the general labor force, of the population as a whole, and of various subgroups of the population. Monthly labor force data for the country are used by the &lt;ahref="http://www.bls.gov"&gt;Bureau of Labor Statistics (BLS)&lt;/a&gt; to determine the distribution of funds under the Job Training Partnership Act. These data are collected through combined computer-assisted personal interviewing (CAPI) and computer-assisted telephone interviewing (CATI). In addition to the labor force data, the CPS basic funding provides annual data on work experience, income, health insurance, and migration data from the Annual Social and Economic Supplement (ASEC), and on school enrollment of the population from the October Supplement. Other supplements, some of which are sponsored by other agencies, are conducted biennially or intermittently.
#> 6 To provide estimates of employment, unemployment, and other characteristics of the general labor force, of the population as a whole, and of various subgroups of the population. Monthly labor force data for the country are used by the &lt;ahref="http://www.bls.gov"&gt;Bureau of Labor Statistics (BLS)&lt;/a&gt; to determine the distribution of funds under the Job Training Partnership Act. These data are collected through combined computer-assisted personal interviewing (CAPI) and computer-assisted telephone interviewing (CATI). In addition to the labor force data, the CPS basic funding provides annual data on work experience, income, health insurance, and migration data from the Annual Social and Economic Supplement (ASEC), and on school enrollment of the population from the October Supplement. Other supplements, some of which are sponsored by other agencies, are conducted biennially or intermittently.
#>              contact
#> 1 dsd.cps@census.gov
#> 2 dsd.cps@census.gov
#> 3 dsd.cps@census.gov
#> 4 dsd.cps@census.gov
#> 5 dsd.cps@census.gov
#> 6 dsd.cps@census.gov