Get variable or geography metadata for a given API as a data frame

Get variable or geography metadata for a given API as a data frame

listCensusMetadata(name, vintage = NULL, type = "variables", group = NULL)

Arguments

name
API name - e.g. acs5. See list at https://api.census.gov/data.html

vintage
Vintage of dataset, e.g. 2014 - not required for timeseries APIs

type
Type of metadata to return, either "variables", "geographies" or "geography", or "groups". Default is variables.

group
An optional variable group code, used to return metadata for a specific group of variables only.

Examples

bds_vars <- listCensusMetadata(name = "timeseries/bds/firms", type = "variables")
head(bds_vars)

#>                      name                                          label
#> 1       reallocation_rate                              Reallocation rate
#> 2  job_destruction_deaths                         Job destruction deaths
#> 3 job_creation_continuers                        Job creation continuers
#> 4                    year                                 Year of survey
#> 5             estabs_exit                          Establishments exited
#> 6                  estabs Physical locations where business is conducted
#>               concept predicateType group limit required
#> 1 Economic Statistics           int   N/A     0     <NA>
#> 2 Economic Statistics           int   N/A     0     <NA>
#> 3 Economic Statistics           int   N/A     0     <NA>
#> 4 Economic Statistics           int   N/A     0     <NA>
#> 5 Economic Statistics           int   N/A     0     <NA>
#> 6 Economic Statistics           int   N/A     0     <NA>

bds_geos <- listCensusMetadata(name = "timeseries/bds/firms", type = "geographies")
head(bds_geos)

#>                            name geoLevelId referenceDate
#> 1                            us        010    2012-01-01
#> 2 metropolitan statistical area        310          <NA>
#> 3                         state        040    2012-01-01

acs_geos <- listCensusMetadata(name = "acs/acs5", vintage = 2017, type = "geographies")
head(acs_geos)

#>                 name geoLevelDisplay referenceDate      requires wildcard
#> 1                 us             010    2017-01-01          NULL     NULL
#> 2             region             020    2017-01-01          NULL     NULL
#> 3           division             030    2017-01-01          NULL     NULL
#> 4              state             040    2017-01-01          NULL     NULL
#> 5             county             050    2017-01-01         state    state
#> 6 county subdivision             060    2017-01-01 state, county   county
#>   optionalWithWCFor
#> 1              <NA>
#> 2              <NA>
#> 3              <NA>
#> 4              <NA>
#> 5             state
#> 6            county

acs_groups <- listCensusMetadata(name = "acs/acs5", vintage = 2017, type = "groups")
head(acs_groups)

#>     name
#> 1 B17015
#> 2 B18104
#> 3 B17016
#> 4 B18105
#> 5 B17017
#> 6 B18106
#>                                                                                                                                                         description
#> 1 POVERTY STATUS IN THE PAST 12 MONTHS OF FAMILIES BY FAMILY TYPE BY SOCIAL SECURITY INCOME BY SUPPLEMENTAL SECURITY INCOME (SSI) AND CASH PUBLIC ASSISTANCE INCOME
#> 2                                                                                                                                SEX BY AGE BY COGNITIVE DIFFICULTY
#> 3                                                      POVERTY STATUS IN THE PAST 12 MONTHS OF FAMILIES BY FAMILY TYPE BY WORK EXPERIENCE OF HOUSEHOLDER AND SPOUSE
#> 4                                                                                                                               SEX BY AGE BY AMBULATORY DIFFICULTY
#> 5                                                                                      POVERTY STATUS IN THE PAST 12 MONTHS BY HOUSEHOLD TYPE BY AGE OF HOUSEHOLDER
#> 6                                                                                                                                SEX BY AGE BY SELF-CARE DIFFICULTY
#>                                                      variables
#> 1 https://api.census.gov/data/2017/acs/acs5/groups/B17015.json
#> 2 https://api.census.gov/data/2017/acs/acs5/groups/B18104.json
#> 3 https://api.census.gov/data/2017/acs/acs5/groups/B17016.json
#> 4 https://api.census.gov/data/2017/acs/acs5/groups/B18105.json
#> 5 https://api.census.gov/data/2017/acs/acs5/groups/B17017.json
#> 6 https://api.census.gov/data/2017/acs/acs5/groups/B18106.json

group_B17020 <- listCensusMetadata(name = "acs/acs5",
vintage = 2017,
type = "variables",
group = "B17020")
head(group_B17020)

#>           name
#> 1 B17020_017EA
#> 2 B17020_016MA
#> 3 B17020_016EA
#> 4 B17020_015MA
#> 5 B17020_015EA
#> 6 B17020_014EA
#>                                                                                                          label
#> 1     Annotation of Estimate!!Total!!Income in the past 12 months at or above poverty level!!85 years and over
#> 2 Annotation of Margin of Error!!Total!!Income in the past 12 months at or above poverty level!!75 to 84 years
#> 3        Annotation of Estimate!!Total!!Income in the past 12 months at or above poverty level!!75 to 84 years
#> 4 Annotation of Margin of Error!!Total!!Income in the past 12 months at or above poverty level!!60 to 74 years
#> 5        Annotation of Estimate!!Total!!Income in the past 12 months at or above poverty level!!60 to 74 years
#> 6        Annotation of Estimate!!Total!!Income in the past 12 months at or above poverty level!!18 to 59 years
#>   predicateType  group limit predicateOnly concept
#> 1        string B17020     0          TRUE    <NA>
#> 2        string B17020     0          TRUE    <NA>
#> 3        string B17020     0          TRUE    <NA>
#> 4        string B17020     0          TRUE    <NA>
#> 5        string B17020     0          TRUE    <NA>
#> 6        string B17020     0          TRUE    <NA>