Get variable or geography metadata for a given API as a data frame

Get variable or geography metadata for a given API as a data frame

listCensusMetadata(name, vintage = NULL, type = "variables")

Arguments

name
API name - e.g. acs5. See list at https://api.census.gov/data.html

vintage
Vintage of dataset, e.g. 2014 - not required for timeseries APIs

type
Type of metadata to return, either "variables", "geographies" or "geography", or "groups". Default is variables.

Examples

bds_vars <- listCensusMetadata(name = "timeseries/bds/firms", type = "variables")
head(bds_vars)

#>                       name                    label                concept
#> 1 job_creation_rate_births Job creation rate births    Economic Statistics
#> 2                     sic1          Industry Sector    Economic Statistics
#> 3        estabs_entry_rate Establishment entry rate    Economic Statistics
#> 4                   ifsize        Initial firm size    Economic Statistics
#> 5              estabs_exit    Establishments exited    Economic Statistics
#> 6                    state          State FIPS Code Selectable Geographies
#>   predicateType group limit          required
#> 1           int   N/A     0              <NA>
#> 2           int   N/A     0 default displayed
#> 3           int   N/A     0              <NA>
#> 4        string   N/A     0 default displayed
#> 5           int   N/A     0              <NA>
#> 6          <NA>   N/A     0              <NA>

bds_geos <- listCensusMetadata(name = "timeseries/bds/firms", type = "geography")
head(bds_geos)

#>                            name geoLevelId referenceDate
#> 1                            us        010    2012-01-01
#> 2 metropolitan statistical area        310          <NA>
#> 3                         state        040    2012-01-01

acs_geos <- listCensusMetadata(name = "acs/acs5", vintage = 2016, type = "geography")
head(acs_geos)

#>                 name geoLevelDisplay referenceDate      requires wildcard
#> 1                 us             010    2015-01-01          NULL     NULL
#> 2             region             020    2015-01-01          NULL     NULL
#> 3           division             030    2015-01-01          NULL     NULL
#> 4              state             040    2015-01-01          NULL     NULL
#> 5             county             050    2015-01-01         state    state
#> 6 county subdivision             060    2015-01-01 state, county   county
#>   optionalWithWCFor
#> 1              <NA>
#> 2              <NA>
#> 3              <NA>
#> 4              <NA>
#> 5             state
#> 6            county

acs_groups <- listCensusMetadata(name = "acs/acs5", vintage = 2016, type = "groups")
head(acs_groups)

#>      name
#> 1  B07012
#> 2  B07013
#> 3 B13002A
#> 4 B13002C
#> 5 B13002B
#> 6 B13002E
#>                                                                                                                       description
#> 1       GEOGRAPHICAL MOBILITY IN THE PAST YEAR BY POVERTY STATUS IN THE PAST 12 MONTHS FOR CURRENT RESIDENCE IN THE UNITED STATES
#> 2                                     GEOGRAPHICAL MOBILITY IN THE PAST YEAR BY TENURE FOR CURRENT RESIDENCE IN THE UNITED STATES
#> 3                                      WOMEN 15 TO 50 YEARS WHO HAD A BIRTH IN THE PAST 12 MONTHS BY MARITAL STATUS (WHITE ALONE)
#> 4          WOMEN 15 TO 50 YEARS WHO HAD A BIRTH IN THE PAST 12 MONTHS BY MARITAL STATUS (AMERICAN INDIAN AND ALASKA NATIVE ALONE)
#> 5                  WOMEN 15 TO 50 YEARS WHO HAD A BIRTH IN THE PAST 12 MONTHS BY MARITAL STATUS (BLACK OR AFRICAN AMERICAN ALONE)
#> 6 WOMEN 15 TO 50 YEARS WHO HAD A BIRTH IN THE PAST 12 MONTHS BY MARITAL STATUS (NATIVE HAWAIIAN AND OTHER PACIFIC ISLANDER ALONE)
#>                                                       variables
#> 1  https://api.census.gov/data/2016/acs/acs5/groups/B07012.json
#> 2  https://api.census.gov/data/2016/acs/acs5/groups/B07013.json
#> 3 https://api.census.gov/data/2016/acs/acs5/groups/B13002A.json
#> 4 https://api.census.gov/data/2016/acs/acs5/groups/B13002C.json
#> 5 https://api.census.gov/data/2016/acs/acs5/groups/B13002B.json
#> 6 https://api.census.gov/data/2016/acs/acs5/groups/B13002E.json