Get variable or geography metadata for a given API as a data frame

Get variable or geography metadata for a given API as a data frame

listCensusMetadata(name, vintage = NULL, type = "variables")

Arguments

name
API name - e.g. acs5. See list at https://api.census.gov/data.html

vintage
Vintage of dataset, e.g. 2014 - not required for timeseries APIs

type
Type of metadata to return, either "variables" or "v" to return variables or "geographies" or "g" to return geographies. Default is variables.

Examples

varsbds<- listCensusMetadata(name="timeseries/bds/firms", type = "v")
head(varsbds)

#>                       name                    label                concept
#> 1 job_creation_rate_births Job creation rate births    Economic Statistics
#> 2                     sic1          Industry Sector    Economic Statistics
#> 3        estabs_entry_rate Establishment entry rate    Economic Statistics
#> 4                   ifsize        Initial firm size    Economic Statistics
#> 5              estabs_exit    Establishments exited    Economic Statistics
#> 6                    state          State FIPS Code Selectable Geographies
#>   predicateType limit required
#> 1           int     0       NA
#> 2           int     0       NA
#> 3           int     0       NA
#> 4        string     0       NA
#> 5           int     0       NA
#> 6          <NA>     0       NA

geosbds <- listCensusMetadata(name="timeseries/bds/firms", type = "g")
head(geosbds)

#>                            name geoLevelId
#> 1                            us        010
#> 2 metropolitan statistical area        310
#> 3                         state        040

geosacs <- listCensusMetadata(name="acs5", vintage = 2015, type = "g")
head(geosacs)

#>                                                          name geoLevelId
#> 1 metropolitan statistical area/micropolitan statistical area        310
#> 2                                                          us        010
#> 3                                                      region        020
#> 4                                                    division        030
#> 5                                                  urban area        400
#> 6  american indian area/alaska native area/hawaiian home land        250
#>   requires wildcard optionalWithWCFor
#> 1     NULL     NULL              <NA>
#> 2     NULL     NULL              <NA>
#> 3     NULL     NULL              <NA>
#> 4     NULL     NULL              <NA>
#> 5     NULL     NULL              <NA>
#> 6     NULL     NULL              <NA>