Use variable metadata to find variables containing a given string.

Return a list of variable names or data frame of variable metadata containing a given string. This can be used create a list of variables to later pass to getCensus, or a data frame documenting variables used in a given project.

makeVarlist(name, vintage = NULL, find, varsearch = "all",
  output = "list")

Arguments

name
API name - e.g. acs5. See list at https://api.census.gov/data.html

vintage
Year of dataset, e.g. 2014 - not required for timeseries APIs

find
A string to find in the variable metadata

varsearch
Optional argument specifying which fields to search. Default is "all". Options are "all", "name", "label", or "concept".

output
Optional argument, specifying output to "list" or "dataframe". Default is "list".

Examples

# Return a list, and then use getCensus function to retrieve those variables

myvars <- makeVarlist(name = "timeseries/poverty/saipe",
  find = "Ages 0-4",
  varsearch = "label")
myvars

#> [1] "SAEPOVRT0_4_UB90" "SAEPOVRT0_4_MOE"  "SAEPOVU_0_4"      "SAEPOVRT0_4_PT"  
#> [5] "SAEPOVRT0_4_LB90" "SAEPOV0_4_UB90"   "SAEPOV0_4_MOE"    "SAEPOV0_4_PT"    
#> [9] "SAEPOV0_4_LB90"  

saipe_dt <- getCensus(name = "timeseries/poverty/saipe",
  time = 2016,
  vars = myvars,
  region = "state:*")
head(saipe_dt)

#>   time state SAEPOVRT0_4_UB90 SAEPOVRT0_4_MOE SAEPOVU_0_4 SAEPOVRT0_4_PT
#> 1 2016    01             28.9             1.5      287177           27.4
#> 2 2016    02             17.0             1.9       52408           15.1
#> 3 2016    04             26.1             1.2      428317           24.9
#> 4 2016    05             29.1             1.8      184115           27.3
#> 5 2016    06             21.2             0.5     2430975           20.7
#> 6 2016    08             15.1             1.2      330815           13.9
#>   SAEPOVRT0_4_LB90 SAEPOV0_4_UB90 SAEPOV0_4_MOE SAEPOV0_4_PT SAEPOV0_4_LB90
#> 1             25.9          83093          4418        78675          74257
#> 2             13.2           8940          1021         7919           6898
#> 3             23.7         111948          5131       106817         101686
#> 4             25.5          53587          3246        50341          47095
#> 5             20.2         515449         13017       502432         489415
#> 6             12.7          50005          3894        46111          42217